+++ Scripting/InedoExtension/ScriptLanguages/Shell/ShellScriptingLanguage.cs
@@ -14,9 +14,9 @@ namespace Inedo.Extensions.Scripting.ScriptLanguages.Shell
         public override string FileExtension => ".sh";
         public override string SyntaxName => "sh";
         public override FileResource Icon { get; } = new EmbeddedFileResource("Shell.bash.svg", "image/svg+xml");
-        public override Type CallOperationType => typeof(SHCallNewOperation);
-        public override Type EnsureOperationType => typeof(SHEnsureNewOperation);
-        public override Type VerifyOperationType => typeof(SHVerifyNewOperation);
+        public override Type CallOperationType => typeof(SHCallOperation);
+        public override Type EnsureOperationType => typeof(SHEnsure2Operation);
+        public override Type VerifyOperationType => typeof(SHVerify2Operation);
         public override ScriptParameterUsage ParameterUsage => ScriptParameterUsage.Arguments | ScriptParameterUsage.InputVariable | ScriptParameterUsage.OutputVariable;
         protected override ScriptInfo ParseScriptInfo(RaftItem2 script)
         {
